refactor: split serial in header and impl.
